+++ NEWS.md
@@ -1,3 +1,8 @@
+1.5.2
+-----
+* Keep numbers as such in JSON output (#100)
+* Revert max-age=0 in Cache-Control when using no-cache
+
 1.5.1
 -----
 * Improve cacheability of queries selecting "updated_at" fields (#99)